Move keypair gen. to prepare() and fix errhandling
The keypair should only be generated if the parameters have been verified and
all other validation has succeeded. Therefore it's better to move it from
OA2ProxyServlet.verifyAndGet() to MPOA2ProxyServlet.prepare() (the latter class
extends the former). Unfortunately there is no easy place to do this in the
myproxy package, only in the master-portal package.
Also fix error handling, whether errors are from the myproxy-info command or
from the input parameters, and making sure that the keypair is only created when
all those are true. For this we also make the validators able to handle a null
value for the myproxy-info result. The problem there is that we have one
validator type for myproxy-info result and request parameter validation.
Fix typo in exception class.