Fix internal server error in case of missing code
Since we use the state parameter for keeping state also between the server and
client inside the MP, we need the state. Check for it in the
MPOA2AuthorizationServer.java and throw a invalid request to the user.
At the client side, if it missed the state and/or code, it returns a 400 to the
server servlet, but that should in itself still return as a internal server
error to the user (which in practise will not occur due to the catching before
the forward).
Also fix a typo.