Add (default disabled) autoregistration endpoint
Add a new autoregistration endpoint which can be used with a POST API call.
All the normal client parameters can be passed in, plus a parameter approverid
which can be e.g. a valid URL which will be added (URLencoded) to the
default approver ("auto-approver") separated with a :.
Succesful registration will return a JSON containing client_id and
client_secret.
The implementation is in MPOA2AutoRegistrationServlet.java

Autoregistration (i.e. without approval) will be necessary if we want to support
OIDC federation using a reverse proxy.
By default the /autoregistration endpoint is disabled, it requires setting
enableAutoRegisterEndpoint="true" in the cfg.xml file.

NOTE: the /autoregistration MUST NOT be exposed to the web.