Update for new API, although not used
Update for the new API, i.e. going from (Basic)ScopeHandler to
(Basic)ClaimsSourceImpl. Also rename file correspondingly.
It was originally used for ensuring non-basic claims where forwarded from the
mp-client to the mp-server (i.e. from the DS to the VO portal), this is done
elsewhere and this filter has no purpose currently. We could decide we want it
at some point, e.g. to filter out certain claims we get from the DS, or add
others.
To enable:
- setEnabled(true) in the constructor
- isRunAtAuthorization() should return true