Cleanup and clarify MPForwardingClaimsSourceImpl
The MPForwardingClaimsSourceImpl is currently not doing anything.
Put lots of clarifying text in the file and remove the overridden methods we
shouldn't override since they aren't (shouldn't be) used in any case.
We should ideally add a getClaims() that gets the supported claims from the DS
but that's non-trivial as it involves the mp-client.