Assume posted data is UTF-8 for SSH key API
Assume UTF-8 for posted data. We need to do this in a overridden doPost() as it
has to be done before any access to the request parameters is made (hence doIt()
is to late).
Note that this only works for POST, not GET. For GET we can only practically do
it using the tomcat config file.